LPS-101678 Extract productMenuSettings
