+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/app/ProductMenuSettings.js
@@ -1,0 +1,115 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import React, {useContext} from 'react';
+import {AppDeploymentContext} from './AppDeploymentContext.es';
+
+const SCOPES = [
+	{
+		label: Liferay.Language.get('control-panel'),
+		value: ['control_panel']
+	},
+	{
+		label: Liferay.Language.get('site-menu'),
+		value: ['site_administration.content']
+	},
+	{
+		label: Liferay.Language.get('control-panel-and-site-menu'),
+		value: ['control_panel', 'site_administration.content']
+	}
+];
+
+export default () => {
+	const {
+		state: {
+			app: {appDeployments}
+		},
+		dispatch
+	} = useContext(AppDeploymentContext);
+
+	const onSettingsChange = event => {
+		const newScope = event.target.value;
+
+		const newSettings = {
+			...appDeployments.find(
+				deployment => deployment.type === 'productMenu'
+			).settings,
+			scope: newScope.split(',')
+		};
+
+		dispatch({
+			deploymentType: 'productMenu',
+			settings: newSettings,
+			type: 'UPDATE_DEPLOYMENT_SETTINGS'
+		});
+	};
+
+	const productMenu = appDeployments.find(
+		deployment => deployment.type === 'productMenu'
+	);
+
+	if (!productMenu) {
+		return <></>;
+	}
+
+	const {
+		settings: {scope}
+	} = productMenu;
+
+	return (
+		<>
+			<div className="autofit-row pl-4 pr-4">
+				<div className="autofit-col-expand">
+					<div className="form-group">
+						<label htmlFor="scope">
+							{Liferay.Language.get('place-it-in-the')}
+						</label>
+						<select
+							className="form-control"
+							id="scope"
+							onChange={onSettingsChange}
+							value={scope}
+						>
+							{SCOPES.map(({label, value}, index) => (
+								<option key={index} value={value}>
+									{label}
+								</option>
+							))}
+						</select>
+					</div>
+				</div>
+
+				{scope.includes('site_administration.content') && (
+					<div className="col-md-6">
+						<div className="form-group">
+							<label htmlFor="site">
+								{Liferay.Language.get('site')}
+							</label>
+							<select
+								className="form-control"
+								disabled={true}
+								id="site"
+								value={1}
+							>
+								<option value={1}>
+									{Liferay.Language.get('all-sites')}
+								</option>
+							</select>
+						</div>
+					</div>
+				)}
+			</div>
+		</>
+	);
+};
